src/requirements.txt
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
idna==3.2
jmespath==0.10.0
mangum==0.11.0
orjson==3.5.2
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn
watchgod==0.7
websockets==8.1